pyproject.toml
[tool.poetry]
name = "pypolygon"
version = "0.1.0"
description = "Python library to interface with trading API service Polygon.io"
authors = ["Simone Sturniolo <simonesturniolo@gmail.com>"]
license = "GPLv3"
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.10,<4"
python-dotenv = "^1.0.1"
requests = "^2.31.0"
pydantic = "^2.6.1"


[tool.poetry.group.dev.dependencies]
pytest = "^8.0.0"
mypy = "^1.8.0"
black = "^24.1.1"
ruff = "^0.2.1"
types-requests = "^2.31.0.20240125"
requests-mock = "^1.11.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
